issues/factory-boy-typing-issue/pyproject.toml
[project]
name = "factory-boy-typing-issue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = []

[dependency-groups]
dev = ["factory-boy>=3.3.3", "pyright>=1.1.408"]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.14"
pythonPlatform = "Linux"
